// src/app/models/materiel.model.ts
export interface Materiel {
  materielId: number;
  materielMarque: string;
  materielRef: string;
  materielQuantite: string;
  locId: number;


  pretID: number;
  pretDescription: string;
  userName: string;
  preDebut: string;
  pretFin: string;
  pretQuantite: string;
  pretValide: string;
  prolongationValide: string;
}